src/salmon_project/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
pillow==11.1.0
python-dotenv==1.0.0
whitenoise==6.7.0
commonmark==0.9.1
drf-yasg==1.21.9
mkdocs==1.6.1
gunicorn==23.0.0
psycopg2==2.9.10
requests==2.32.3
django-cors-headers==4.7.0